refactor(tokens): removing tokens/scrollbar
